move modal to separate component
